Import expression tests from pypy (#6)
What helped a ton at the beginning of prettier was to import all the tests in the flow codebase. This shows has a ton of edge cases around JavaScript. I haven't found a similar test suite yet but this file from pypy has a bunch of interesting ones.

I've commented out all the tests that either fatal or output a different result.