Add support for comments
This wires up the comment logic from prettier. This needs some modification inside of prettier itself, so it won't work as is.

- locStart/End is ast-specific, we need a way to set it up for each parser. In this case I reused on that already existed: https://github.com/prettier/prettier/blob/master/src/util.js#L260-L262

- getSortedChildNodes as well, need a way to traverse the ast: https://github.com/prettier/prettier/blob/master/src/comments.js#L38

```js
        (node.kind && node.kind !== "Comment") ||
        node.ast_type && node.ast_type !== "comment") // <-- added
    ) {
```

- printComments as well: https://github.com/prettier/prettier/blob/master/src/comments.js#L928-L929

```js
  switch (comment.type || comment.kind || comment.ast_type) {
    case "comment":
      return comment.value;
    case "Comment":
```

If we tweak those two small things in prettier itself, then it's able to print comments using the heuristic. @azz, do you think you'd be able to extract those out to be parser-specific?