Upgrade Prettier version, fix comment discovery
Bump the dependency on the main Prettier repo to the latest commit. This allows us to specify additional options (such as `pythonVersion`). Note that the version is defaulting to Python3, which I believe matches the behavior from before.

The plugin API in the newer version requires plugins to provide a `locStart` and `locEnd` function. From what I've seen, this is to allow customization, which is something we want and will help with things like better comment rendering. However, I'm not totally sure what customizations we want/need in these functions, so for now this just copies in the old versions of those functions. This shouldn't actually be a functional change from before the upgrade, and still allows us to modify/customize that functionality in the future as needed.

I also noticed a bug in how comments are detected within Python. There was a check that hard-coded the `COMMENT` token type as `57`, but it's actually different on different versions of Python. This fixes that check to use the proper value for the given version of Python and updates a test that mistakenly did not have comments in the expected output.