Fail tests when encountering unknown node types
Before, a basic test added for a new node type would not fail, since the default behavior for a node would be sufficient. While more advanced tests *would* fail (because the desired formatting behavior didn't happen), it still seems useful to have tests assert that all seen node types are known.

I tested this by adding a test for a new node type (in a different diff) and seeing that the test failed as expected after making this change. I ran `yarn prettier` on the test file and saw that, while it complained about an unknown node, it didn't fail.