Clean up handling of list-like literals (#28)
This makes list-like literals consistently have the following behavior:
- If the list is too long, it will split items onto their own lines, each indented.
- If the option is set, the list item will have a trailing comma when split onto its own line.

This also does some cleanup of Tuples to be a bit more consistent with best practices around when to have/not have parens (as I understand them). I realize this is a bit contentious and I'm certainly open to changing/reverting specific behaviors here. In general, the way I've coded in the past is to skip parens when on a single line and either unpacking into multiple variables or returning from a function, and everything else has parens. Again, up for debate :)

While I'm here, fix multi-assign (assignments with multiple equals signs) to be able to test how that interacts with tuples.