Don't try to canonicalize strings in multiple parts (#32)
We currently try to canonicalize strings to use consistent quotes. However, there are a number of cases where this breaks and produces invalid Python code. As a short term stop-gap, this diff tries to detect those scenarios and bails early, leaving the string as-is.

This diff adds a number of tests which expose this behavior. Some of the tests are set up to maybe one day check more advanced behavior (like automatically re-flowing strings) but for now they just test the fixes here. There are also a few new tests commented out because we still don't handle them properly, and either crash or produce invalid Python. This also adds some comments for similar tests we could add for f-strings, but those require a fix in the ASTTokens library to work correctly.