Update in-repo copy of asttokens (#41)
This updates the code to 3518f5ae323f5a390f8894557db8285eb1ec7e34. In the future we might want to look into using a virtualenv or something like that so we don't need to keep our own copy of this code, but this isn't any worse than the existing situation.