More intuitive same-line placement for list-like values
This diff tries to add more intuitive/reasonable placement for values in certain situations. In particular, if we have list-like literals that are too long to fit on a line, we often want the open/close brackets of the literal to be on the same line as the surrounding structure, rather than be entirely on their own line. An example is:

```
foo([
    1,
    2,
    ...
    10,
])
```

The naive way of using `group`s would cause the `[` and `]` to each be on their own lines, due to how `group` splits from the outside in. This adds optimizations/special cases for this style of formatting in `Assign`s (and the closely related `AnnAssign`s and `AugAssign`s), `Call`s, `Return`s and `Yield`s.

There are fairly extensive tests for the new behavior, but let me know if there's any other cases you'd like to see! In particular there's tests for non-list-like expressions to make sure that "normal" wrapping works as indended, and tests for the various kinds of assignments. There's still some work for how we handle complex (multi-part) strings, but I think that's mostly unrelated to this diff.

I also tried to comment it so it'd be easier to read through in the future, but let me know if there's any part that is confusing or could have better comments!